Update versionhistory.html
HOTFIX - Simple change, removing the / before the links. This was a basic error preventing the live site to run the correct paths, breaking the structure of the page.
This is a logical decision and after following the correct PR procedure last time out, I have got to add an extra change to the file.