v2.6.7
Added a landing page to the site, whilst I do a big revamp in the background that I am working on.

The aim here is to have a fully functional portfolio, without my website being live as I am adding all of my work in and making design / interactivity tweaks.