v2.70
Amended the version history page with the landing page that was online whilst I went from v2.5 to v3.0. This landing page consists of some of my CodePen projects and social / contact links.

I am revamping the site design and updating content. It's quite a big operation and need to streamline operations also.