V2.8
Added sitemap page, styled and added this into the website footer on main page.
Added pages to update inside the sitemap for ease of reference in the branch, it will gradually be updated.

Removed thank you page.

Merged in v2.7 into this branch for consistency as working on the main site, not the landing page (which is current main which I branched off to create 2.8 and 2.7 branches)